Longer timeout for grpc stream/websocket
subscribeInvoice() grpc stream endpoint is cut off after 1 min or so. I guessed it is because of the nginx default timeout.
This change sets 1 day for long connection. I have tested with node.js grpc client and at least the connection is not forced to get cut off after this change.
I am not so sure whether 1 day is too long or too short. or even  whether BTCPay does not need to expose it.