Add a simple template mechanism for fragments
This small change extends the generator to be able to replace placeholder test in fragments. Right now I'm only using it for an internal value (Cryptos, replaces #CRYPTOS# in the generated yml with the selected crypto codes e.g.: "BTC;LTC") but it could prove to be useful for other things later on.