Integrate EPS (Electrum Personal Server)
After the ElectrumX integration in the last release, I thought that some people may prefer the Electrum Personal Server (EPS) approach, which can only serve specific pre-determined wallets (using XPUBS).  EPS cannot serve as a server for ad-hoc wallets with connections for many people, that ElectrumX delivers.  The upside pf EPS is that it is much lighter, and doesn't require txindex=1.

I will integrate it with this pull request, but it might take a while, as I have some busy weeks ahead of me.