change line to . ./btcpay-setup.sh -I line  177
both Method works
.
. .
but this is the way describe in btcpay tooling scripts .

extra info for newbies
On Unix-like operating systems every directory contains, as a minimum, an object represented by a single dot and another represented by two successive dots. The former refers to the directory itself and the latter refers to its parent directory (i.e., the directory that contains it).